requirements.txt
git+https://github.com/paulordyl/DXcam.git
cityhash>=0.4.7,<0.5.0
mypy>=1.1.1,<2.0.0
nptyping>=2.5.0,<3.0.0
numba>=0.58.1,<0.59.0
numpy==1.26.2
opencv-python>=4.6.0.66,<5.0.0.0
pillow>=9.1.1,<10.0.0
pyautogui>=0.9.53,<0.10.0
pygetwindow>=0.0.9,<0.1.0
pywin32>=306,<307
scipy>=1.9.1,<2.0.0
tcod>=13.8.1,<14.0.0
tinydb>=4.8.0,<5.0.0
customtkinter>=5.2.2,<6.0.0
requests>=2.31.0,<3.0.0
pyserial>=3.5,<4.0
scikit-image>=0.22.0,<0.23.0
pytesseract>=0.3.10,<0.4.0
psutil>=5.9.0,<6.0.0
